Move quick pulse uploader into separate file
