+++ src/uploader_quick_pulse.rs
@@ -1,0 +1,137 @@
+use crate::{models::QuickPulseEnvelope, uploader::serialize_request_body, Error, HttpClient};
+use http::{HeaderName, Request, Uri};
+use std::convert::TryFrom;
+
+// Allow interior mutability. See https://github.com/hyperium/http/issues/599
+#[allow(clippy::declare_interior_mutable_const)]
+const QPS_TRANSMISSION_TIME: HeaderName = HeaderName::from_static("x-ms-qps-transmission-time");
+#[allow(clippy::declare_interior_mutable_const)]
+const QPS_STREAM_ID: HeaderName = HeaderName::from_static("x-ms-qps-stream-id");
+#[allow(clippy::declare_interior_mutable_const)]
+const QPS_MACHINE_NAME: HeaderName = HeaderName::from_static("x-ms-qps-machine-name");
+#[allow(clippy::declare_interior_mutable_const)]
+const QPS_INSTANCE_NAME: HeaderName = HeaderName::from_static("x-ms-qps-instance-name");
+#[allow(clippy::declare_interior_mutable_const)]
+const QPS_ROLE_NAME: HeaderName = HeaderName::from_static("x-ms-qps-role-name");
+#[allow(clippy::declare_interior_mutable_const)]
+const QPS_INVARIANT_VERSION: HeaderName = HeaderName::from_static("x-ms-qps-invariant-version");
+#[allow(clippy::declare_interior_mutable_const)]
+const QPS_SUBSCRIBED: HeaderName = HeaderName::from_static("x-ms-qps-subscribed");
+#[allow(clippy::declare_interior_mutable_const)]
+const QPS_REDIRECT: HeaderName = HeaderName::from_static("x-ms-qps-service-endpoint-redirect-v2");
+#[allow(clippy::declare_interior_mutable_const)]
+const QPS_INTERVAL_HINT: HeaderName =
+    HeaderName::from_static("x-ms-qps-service-endpoint-interval-hint");
+
+pub(crate) enum PostOrPing {
+    Post,
+    Ping,
+}
+
+impl std::fmt::Display for PostOrPing {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str(match self {
+            PostOrPing::Post => "post",
+            PostOrPing::Ping => "ping",
+        })
+    }
+}
+
+pub(crate) struct QuickPulseResponse {
+    pub(crate) should_post: bool,
+    pub(crate) redirected_host: Option<http::Uri>,
+    pub(crate) polling_interval_hint: Option<std::time::Duration>,
+}
+
+pub(crate) async fn send(
+    client: &dyn HttpClient,
+    endpoint: &Uri,
+    instrumentation_key: &str,
+    post_or_ping: PostOrPing,
+    envelope: QuickPulseEnvelope,
+) -> Result<QuickPulseResponse, Error> {
+    let endpoint = format!(
+        "{}/QuickPulseService.svc/{}?ikey={}",
+        endpoint, post_or_ping, instrumentation_key
+    );
+
+    let payload = serialize_envelope(&envelope, &post_or_ping)?;
+
+    let mut request_builder = Request::post(endpoint)
+        .header(http::header::EXPECT, "100-continue")
+        .header(QPS_TRANSMISSION_TIME, quick_pulse_transmission_time())
+        .header(http::header::CONTENT_TYPE, "application/json")
+        .header(http::header::CONTENT_ENCODING, "gzip");
+    if matches!(post_or_ping, PostOrPing::Ping) {
+        request_builder = request_builder
+            .header(QPS_STREAM_ID, envelope.stream_id)
+            .header(QPS_MACHINE_NAME, envelope.machine_name)
+            .header(QPS_INSTANCE_NAME, envelope.instance)
+            .header(QPS_INVARIANT_VERSION, envelope.invariant_version);
+        if let Some(role_name) = envelope.role_name {
+            request_builder = request_builder.header(QPS_ROLE_NAME, role_name);
+        }
+    }
+
+    let request = request_builder
+        .body(payload)
+        .expect("request should be valid");
+
+    let response = client
+        .send(request)
+        .await
+        .map_err(Error::UploadConnection)?;
+
+    if response.status().is_success() {
+        let should_post = response
+            .headers()
+            .get(QPS_SUBSCRIBED)
+            .and_then(|v| v.to_str().ok())
+            .map(|v| v == "true")
+            .unwrap_or(false);
+        let redirected_host = response
+            .headers()
+            .get(QPS_REDIRECT)
+            .and_then(|v| v.to_str().ok())
+            .and_then(|v| http::Uri::try_from(v).ok());
+        let polling_interval_hint = response
+            .headers()
+            .get(QPS_INTERVAL_HINT)
+            .and_then(|v| v.to_str().ok())
+            .and_then(|v| v.parse::<u64>().ok())
+            .map(std::time::Duration::from_millis);
+        Ok(QuickPulseResponse {
+            should_post,
+            redirected_host,
+            polling_interval_hint,
+        })
+    } else {
+        Err(Error::Upload(String::new()))
+    }
+}
+
+fn serialize_envelope(
+    envelope: &QuickPulseEnvelope,
+    post_or_ping: &PostOrPing,
+) -> Result<Vec<u8>, Error> {
+    let serialized = match post_or_ping {
+        PostOrPing::Post => serde_json::to_vec(&[&envelope]),
+        PostOrPing::Ping => serde_json::to_vec(&envelope),
+    }
+    .map_err(Error::UploadSerializeRequest)?;
+    serialize_request_body(serialized)
+}
+
+/// UTC time the request was made. Expressed as the number of 100-nanosecond intervals that have
+/// elapsed since 12:00:00 midnight on January 1, 0001. This is used for clock skew calculations,
+/// so the value can never be stale (cached).
+fn quick_pulse_transmission_time() -> String {
+    use std::time::SystemTime;
+
+    let ms_since_0001 = 62135596800000u128;
+    let ms_since_epoch = SystemTime::now()
+        .duration_since(SystemTime::UNIX_EPOCH)
+        .map(|d| d.as_millis())
+        .unwrap_or(0);
+    ((ms_since_0001 + ms_since_epoch) * 10_000).to_string()
+}
